Conflicts:
	cocos/2d/CCTMXXMLParser.cpp
	cocos/2d/CCTMXXMLParser.h

+++ samples/Cpp/TestCpp/Classes/ExtensionsTest/TableViewTest/TableViewTestScene.h
@@ -20,7 +20,7 @@ public:
     virtual void tableCellTouched(cocos2d::extension::TableView* table, cocos2d::extension::TableViewCell* cell);
     virtual cocos2d::Size tableCellSizeForIndex(cocos2d::extension::TableView *table, ssize_t idx);
     virtual cocos2d::extension::TableViewCell* tableCellAtIndex(cocos2d::extension::TableView *table, ssize_t idx);
-    virtual long numberOfCellsInTableView(cocos2d::extension::TableView *table);
+    virtual ssize_t numberOfCellsInTableView(cocos2d::extension::TableView *table);
 };
 
 #endif // __TABLEVIEWTESTSCENE_H__
